Crate rtthw

Table of Contents

rtthw.github.io

What’s going on here?

How to read this site

Modules

Guide to Rust

Blog

rtthw.github.io

The personal website for Matthew Martin.

What’s going on here?

This is an auto-generated website created using rustdoc, and hosted for free through GitHub Pages.

I chose to do it this way because I wanted to create a website for myself, but I am not a web developer. I can’t stand JavaScript, tolerate HTML, and resent CSS. With this method, I can at least sleep soundly at night knowing I’m not a web developer.

How I did it.

How to read this site.

Modules

guide_to_rust

Guide to Rust

posts

Posts